capabilities: add a test file for the new capabilities module, and move the corresponding tests from main to there
