+++ capa/features/capabilities/common.py
@@ -1,3 +1,11 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2023 Mandiant, Inc. All Rights Reserved.
+# Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at: [package root]/LICENSE.txt
+# Unless required by applicable law or agreed to in writing, software distributed under the License
+#  is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and limitations under the License.
 import logging
 import itertools
 import collections
@@ -45,5 +53,5 @@ def find_capabilities(
         return find_static_capabilities(ruleset, extractor, disable_progress=disable_progress, **kwargs)
     if isinstance(extractor, DynamicFeatureExtractor):
         return find_dynamic_capabilities(ruleset, extractor, disable_progress=disable_progress, **kwargs)
-    else:
-        raise ValueError(f"unexpected extractor type: {extractor.__class__.__name__}")
+
+    raise ValueError(f"unexpected extractor type: {extractor.__class__.__name__}")
